ajout d'utilisateurs et roles

+++ backend/src/main/java/edutech/payload/response/JwtResponse.java
@@ -5,13 +5,13 @@ import java.util.List;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private int id;
     private String first_name;
     private String last_name;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, Long id, String first_name, String last_name, String email,
+    public JwtResponse(String accessToken, int id, String first_name, String last_name, String email,
             List<String> roles) {
         this.token = accessToken;
         this.id = id;
@@ -37,11 +37,11 @@ public class JwtResponse {
         this.type = tokenType;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
